pyproject.toml
[project]
name = "memagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.0",
    "langchain-community>=0.4",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=2.0.0",
    "faiss-cpu>=1.7.0",
    "rank-bm25>=0.2.2",
    "python-dotenv>=1.0.0",
    "langchain-google-genai>=2.0.5",
    "pyyaml>=6.0.0",
]
